Reduced var size from uint16 to uint8 in nmea_msgs_t to address compiler warnings and ensure consistentcy in broadcast mechanism

+++ src/data_sets.h
@@ -1431,46 +1431,46 @@ typedef struct PACKED
     uint32_t				options;
 
 	/** Broadcast period multiple - NMEA IMU data. 0 to disable. */
-	uint16_t				pimu;
+	uint8_t				    pimu;
 
 	/** Broadcast period multiple - NMEA preintegrated IMU: delta theta (rad) and delta velocity (m/s). 0 to disable. */
-	uint16_t				ppimu;
+	uint8_t			    	ppimu;
 	
 	/** Broadcast period multiple - NMEA INS output: euler rotation w/ respect to NED, NED position from reference LLA. 0 to disable. */
-	uint16_t				pins1;
+	uint8_t			    	pins1;
 
 	/** Broadcast period multiple - NMEA INS output: quaternion rotation w/ respect to NED, ellipsoid altitude. 0 to disable. */
-	uint16_t				pins2;
+	uint8_t				    pins2;
 	
 	/** Broadcast period multiple - NMEA GPS position data. 0 to disable. */
-	uint16_t				pgpsp;
+	uint8_t			    	pgpsp;
 
 	/** Broadcast period multiple - NMEA Raw IMU data (up to 1KHz).  Use this IMU data for output data rates faster than DID_FLASH_CONFIG.startupNavDtMs.  Otherwise we recommend use of pimu or ppimu as they are oversampled and contain less noise. 0 to disable. */
-	uint16_t				primu;
+	uint8_t			    	primu;
 
 	/** Broadcast period multiple - NMEA standard GGA GNSS 3D location, fix, and accuracy. 0 to disable. */
-	uint16_t				gga;
+	uint8_t				    gga;
 
 	/** Broadcast period multiple - NMEA standard GLL GNSS 2D location and time. 0 to disable. */
-	uint16_t				gll;
+	uint8_t			    	gll;
 
 	/** Broadcast period multiple - NMEA standard GSA GNSS DOP and active satellites. 0 to disable. */
-	uint16_t				gsa;
+	uint8_t			    	gsa;
 
 	/** Broadcast period multiple - NMEA standard recommended minimum specific GPS/Transit data. 0 to disable. */
-	uint16_t				rmc;
+	uint8_t			    	rmc;
 	
 	/** Broadcast period multiple - NMEA standard Data and Time. 0 to disable. */
-	uint16_t				zda;
+	uint8_t			    	zda;
 
 	/** Broadcast period multiple - NMEA standard Inertial Attitude Data. 0 to disable. */
-	uint16_t				pashr;
+	uint8_t			    	pashr;
 
 	/** Broadcast period multiple - NMEA standard satelliate information. */
-	uint16_t				gsv;
+	uint8_t			    	gsv;
 
 	/** Broadcast period multiple - NMEA track made good and speed over ground. */
-	uint16_t				vtg;
+	uint8_t			    	vtg;
 
 } nmea_msgs_t;
